Handle magic calls to unknown methods in interpreter
Summary:
Magic `__call` calls are deprecated and sparsely used. Do not emit a ton of
code to handle them when emitting an unknown method call. Instead, go back to
the interpreter and deal with them there.

Since magic calls will be no longer profiled, we also don't need to handle them
in optimizeProfiledCallMethod().

Reviewed By: ricklavoie

Differential Revision: D17083706

fbshipit-source-id: b2a4479dcbc7bb5dc22fb04cce764ffcc25648ef