Handle magic calls to known methods at the caller
Summary:
Call the `__call()` directly, passing a method name and a vec/varray of the
arguments constructed by the caller instead of setting m_invName on the ActRec
and relying on the prologue to do the work.

Let the inliner inline the call, since there is nothing special about it.

Reviewed By: ricklavoie

Differential Revision: D17085128

fbshipit-source-id: a1867a02d484dfe534fc2c14dbc55121fb061ac0