Fix gating for simple pessimization
Summary:
General summary: this is meant to deterministically decide whether to turn on simple pessimization for a portion of the Hack codebase.

Previously, the hash of the filename was much smaller than Int.max_value, so when divided the result was `0.000000...`. The effect was that even `simple_pessimize = 0.001` would turn on pessimization for all the files. Since this is just for testing, we don't have to be too concerned with modulo bias. This uses a mod to normalize the hashes down to a useful range before getting a ratio.

Reviewed By: Matt-Schellhas

Differential Revision: D17059016

fbshipit-source-id: aca00b08c74b6b8d67a451279fecd5c8d639e2e1