New flag for "gather function details"
Summary:
For extremely large repositories of Hack code, it's dangerous to enable lookup of function signatures.  Looking up function signatures in autocomplete can consume many gigabytes worth of memory for such a codebase.

However, there may be some circumstances where we want to enable this feature, such as:
* In single-file execution mode, when you use `hh_single_type_check` or other similar programs to examine only one file.
* In integration tests, where we want the extra detail provided by function signatures to ensure that our program found the right symbol.
* When working on small codebases, or when extra detail is preferred and the memory usage of decl caching is deemed worth it

This diff provides a flag that can be set to enable function signature lookup.  By default this flag is disabled.  Future diffs may provide a mechanism for enabling this flag.

Reviewed By: jewelpit

Differential Revision: D17059234

fbshipit-source-id: a5eb4a0dcc3b330f1c827762307519d86087b326