Enable all-local-changes index mode
Summary:
This diff is part of the stack to enable the open source build to move to the new autocomplete process.

Some of our integration tests work using the `hh_single_type_check` program.  This program doesn't need a global index; it only needs an index for data found within the single file being examined.

To solve this, we will enable a "locally captured files only" mode.  A subsequent diff in this stack will then modify hh_single_type_check so that it will pass any files to this mode whenever executed, to enable all tests to run correctly on that file.

When in LocalIndex mode:
* Only local results will be returned; no results from sqlite or glean or any other custom service will be considered.
* Any calls to update_files or remove_files will modify the local results store.

Reviewed By: jewelpit

Differential Revision: D16929816

fbshipit-source-id: d45f066898672a6e6b19a2f2ad1bd97cf458a678