Port syntax tree rewriter from OCaml to Rust
Summary: Pre- and post-traversal rewriter port from OCaml to Rust. It's not a 1-to-1 port because of the borrow checker, of course. In order to avoid excessive use of clones to satisfy the borrow checker (with a potential performance benefit), I introduced a `Syntax::drain_children` call that relies on `Syntax::fold_over_children_owned`, a new generated method I wrote in `generate_full_fidelity.ml`.

Reviewed By: dabek

Differential Revision: D16828808

fbshipit-source-id: 0f350814df381c5e250b40943ba1f068ec7f961b