Refactor tyvar_info to prepare for global inference
Summary:
For global inference we are gonna distinguish between "global tyvars" and "local tyvars". The difference between the two is that global tyvars must remain unsolved until the end of the typechecking process.
To do that we are updating the tyvar_info type. It can either be a LocalTyvar -- which is treated exactly as before -- or a GlobalTyvar. For now GlobalTyvar are not treated and will in the future be used to indicate that the referenced tyvar is "global" and can be found in a "global tvenv".

Reviewed By: CatherineGasnier

Differential Revision: D16962000

fbshipit-source-id: 4158df0d57ef5ac0fea9c7a3d3bec7e7daffdacb