Make `subtyping_method` interact with `env`
Summary:
The goal of global inference is to widen the scope of the type inferrence: we not want to only infer types locally in a function but globally across functions boundaries. This means that new subtyping relations can be introduced when analyzing methods that overrides an inherited methods. In those cases we want to store those constraints in the env and to be able to use the modified env afterward. This diff:
- Allows an `env` to be returned by functions checking the validity of a class extension (`typing_extends.ml`)
- Allows props to be added to the env in `subtype_method`

Reviewed By: CatherineGasnier

Differential Revision: D17091416

fbshipit-source-id: aa528852c145a37f73c1368d756f2e1cb9364e34