Add "pointer to Rust syntax tree" to OCaml syntax tree
Summary:
Problem: a lot of code is structured as in:

```
1) parse file into syntax tree
2) do a whole lot of other stuff
3) pass the syntax tree into some method
```

We have ported 1) to Rust and many of the things we do (in my case, parser errors) are an example of 3). The best case would be to have all three ported to Rust at once, but in the meantime we could restructure code so it does 1 and 3 before 2 (so that 1 and 3 can be a single Rust method). I tried it and it results in spooky methods like `parse_positioned_with_coroutine_sc_and_rewrite_ppl_classes_and_also_check_errors` and rather unnatural program flows.

The other solution, that I implement here is to be able to leak Rust struct into OCaml and later pass it back to 3). This way minimal changes are necessary, at the cost of introducing a lot of unsafety and possibility for memory leaks. I would prefer not to do this in real production code (though we'll see...), but for the purpose of tests it seems reasonable.

It would be possible to register such pointer with OCaml GC and have it automatically collected, but I think we should not depend on it and be rather precise about what we need when. Not adding GC will just make the memory leak, if it happens, more obvious instead of possibly being obfuscated as unexpectedly high (but bounded) memory consumption that will be much harder to notice or debug.

Reviewed By: shiqicao

Differential Revision: D17107194

fbshipit-source-id: 1fa5e5d4cbff1cdde170776e8855ba861884d2b2