Property redeclaration is covariant when constant
Summary:
# Checking Ambiguous Inheritance #
When checking ambiguous inheritance for non-const properties, class_elts in `typing_defs` cannot currently tell if they have default values. If they don't, all we need to do if check types are equal and there is no ambiguous inheritance. If there are default values, we would need to check that they are the same. I'm unaware of any way to currently do this. The runtime currently checks this.

If the properties are const, we know they MUST have default values, and can error on ambiguous inheritance.

*RFC next step: I can thread a boolean field "has_default_value" through to class_elts to better match runtime behavior of ambiguous inheritance for non-const properties. IIUC booleans are pretty cheap, so this should be fine. I can stack that diff on top of this one. I think it would be quite challenging to prove two values are equal (alias analysis, etc.) and having a boolean should catch most cases*

If property and const, check:
1. If you are abstract and your parent is concrete -> *abstract override concrete error*
2. If you and your parent are concrete, your parent is an interface or trait, and you are not defined in the child class -> *ambiguous inheritance error*
3. Redeclaration is not a subtype of parent declaration -> *wrong type error*

```
abstract class A {
    <<__Const>> abstract public static arraykey $p;
}

class A {
    <<__Const>> public static int $p = 5;
}

```

one_more_try

Reviewed By: kmeht

Differential Revision: D17126571

fbshipit-source-id: 288be2fadd8f5a609b1985c879b8586f090c67ff