Negate the meaning of asyncNonEagerReturnFlag
Summary:
Use zero aux for regular return (i.e. if the function is async, Awaitables are
returned with zero). Use minus one only for special behavior (eager execution).

Always zero-extend the type of return value, even if we are not in a function
that supports async eager return. The type is mostly known anyway, so in most
cases there's no actual zero extend emitted.

Makes the next diff possible, which avoids runtime check for support of async
eager return, causing less code to be emitted when awaiting unknown callees.
Also, it makes it easier to change the behavior of the Call opcode.

Reviewed By: ricklavoie

Differential Revision: D17106528

fbshipit-source-id: be708120652ee2396a04e9906f7f0153634f540e