Do not check for support of async eager return
Summary:
It is okay to set AsyncEagerRet flag when calling a function that does not
support async eager return, as such function would return with zero aux, which
is consistent with not using async eager return.

So, don't bother to check for support of async eager return and unconditionally
check the aux of return value.

Reviewed By: ricklavoie

Differential Revision: D17106698

fbshipit-source-id: 0633b3a86a074d922de95f18d65b90692cc609f8