Implement constant hoisting peephole in vasm-graph-color
Summary:
Implement a simple constant hoisting peephole in vasm-graph-color. Due
to reasons, occasionally the constant placement dataflow will
materialize the same constant at the top of every block which shares
the same predecessor. In this situation, its better to materialize the
constant at the end of the common predecessor. Its difficult to modify
the dataflow to get it to do this, so fixup after the fact with a
quick peephole optimization. Look for blocks where all of its
successors materialize the same constant, and hoist it up to the
parent block.

Reviewed By: ottoni

Differential Revision: D17135337

fbshipit-source-id: d38afcdca895e1edc8a60a1510f6e2b94e9ac16c