Fix crashes when casting TFunc/TClsMeth/TRecord to arrayish types
Summary:
The code was never updated to handle these types, so we'd
assume it was an invalid type (and therefore unreachable), so we'd
crash when we run it. Punt instead until proper behavior can be
implemented. In addition, just punt for unknown types rather than
crashing, which is more safe.

Reviewed By: ottoni, yujunglo

Differential Revision: D17139069

fbshipit-source-id: 7314086ba84a7c33f9f7f358b4458f1c6fda6467