Regenerate oxidized types
Summary: Run the regenerate script `hphp/hack/src/oxidized/regen.sh`. Add a manual definition for `TanySentinel`. Rename the phase type variable in `Tpu_access` from `'a` to `'phase` (so that it is omitted by hh_oxidize). Move the definition of `Tpu_access` from the "Cannot Be Declared In User Code" section to the "Exist in Both Phases" section.

Reviewed By: Wilfred

Differential Revision: D17145252

fbshipit-source-id: 225414dc232fd6c9c04fd271570fd40a1cef7b1e