Profile the type of the output of Shapes::idx
Summary:
We're emitting unnecessary unboxing checks for Shapes::idx results. That's because we don't create a TypeProfile for the result pre-unboxing for Shapes::idx, like we do for ArrayIdx.

The fix is easy; the only thing we have take care with is to clean up after the optimized builtin call when we side exit.

Reviewed By: ottoni

Differential Revision: D17139364

fbshipit-source-id: 08464f93419735206ebf9e6a2ea6f5cfac19ed66