When decls can be deferred, always type check in parallel
Summary:
The deferred decl experiment revealed a bug in the implementation of deferring decls in the type check service: when the number of files to type check is less than 10, we don't use MultiWorker, and process files exactly once. If the files are deferred, we just simply end up not type checking them at all.

Here's a query showing the diffs where the number of errors differs between test and control:
https://fburl.com/scuba/tr9vj3wy

This change makes it so that we always process files in parallel if the decl deferment threshold is set. This makes sense regardless of the bug, too, since a single file may have many decls it depends on.

Reviewed By: Matt-Schellhas

Differential Revision: D17145057

fbshipit-source-id: 6dac357340767df21d3ac906aa0aaf19c7d86f28