Sink IncRef through trivial Jmps
Summary:
Pushing IncRefs past CheckTypes helps us annihilate them with DecRefs, because it can eliminate paths in the CFG where a value is inc-ref-ed but not dec-ref-ed.

In debugging specialized iterator codegen, I found that consecutive CheckTypes are sometimes linked by trivial Jmps (that is, to a block w/ a unique predecessor), because they come from different translations. We don't eliminate these Jmps until later in irgen. We can easily push IncRefs through such Jmps.

This situation is less common outside of specialized iterators, but it still happens and reducing it can be a win. The fix here is a really trivial tweak. I think there's room to be much more aggressive about sinking inc-refs, e.g. being willing to duplicate them in almost all cases where one of the two new ones is in a side-exit block.

Reviewed By: ottoni, mofarrell

Differential Revision: D17114473

fbshipit-source-id: 1708ec15d159007bcd53ba3f5a6a4baf79459af2