Prefetch on VFS-backed repo in typing if decls are deferred
Summary: If we defer declarations, we can also try to prefetch those files that need to be declared. This change is to enable using the VFS prefetch capability when type checking VFS-backed repos. It is meant to be used initially in the context of distributing type checking to remote workers (hulk).

Differential Revision: D17177970

fbshipit-source-id: e123abada23d6817c63689a3c217f18b086a67ea