Make default Eden threshold value optional
Summary:
Just to remove some confusion (mainly, my own), I'm making the remote scheduler env's Eden threshold an optional type, to distinguish it from being set to a specific value by the caller.

In the remote scheduler, the threshold is used to determine if the remote worker's checkout should be backed by Eden or not. It's backed by Eden if the number of files to type check is fewer than the threshold.

Note that I flipped the comparison and it's not identical, but it's more correct now, anyway (the behavior changes AT the threshold, not above it).

Differential Revision: D17186170

fbshipit-source-id: 123c940b02ae3d94f79ac1d431150352662fdb73