Remove broken support for `$x::foo<T>()` reified generics
Summary:
This pattern is the last place that uses ReifiedName opcode, but
FCallClsMethod's JIT implementation never had any support for reified generics,
resulting in inconsistent behavior between interpreter and JIT.

The newly added test demonstrates that. It would previously print "hi" in the
interpreter mode, but throw an exception in JIT mode.

Let's remove its support altogether as it can't be possibly used. Note that the
calls to non-reified method would work fine, so let's just do that.

This diff makes it possible to nuke support for reified names, which allows
merging FCall*RD opcodes into FCall*Ds and their simplification and
generalization, which allows adding support to the FCallClsMethod opcode, which
allows reintroducing actually working support for `$x::foo<T>()`.

Another level of sadness here is inconsistency of stacktraces between
interpreter and JIT. One throws the exception at the callee, while another at
the caller, so I had to use try/catch in the test to produce a consistent
behavior. This is also something that will be fixed once reified generics are
generalized.

Reviewed By: oulgen

Differential Revision: D17181910

fbshipit-source-id: a0b81675e9921c78b2ea49b96eecde59d98e92c9