Fix constant folding of builtins with in-out parameters
Summary:
If a builtin is marked __IsFoldable, both HHBBC and the JIT will try
to const fold any calls to the builtin (if all the parameters are
constants). However this logic failed to deal with builtins which have
in-out parameters properly. Both HHBBC and the JIT would simply const
fold the builtin to the tuple return value (which contains the main
return value and all of the out values). Not only does this lead to
wrong values, but it can cause the stack to become unbalanced (because
the out parameter place holders are not removed properly).

Add the logic to deal with this properly. When const folding builtins
with in-out parameters in the JIT, load the values out of the tuple
and store them to their proper place in the stack. In HHBBC, lower the
FCallBuiltin call to the appropriate series of pops and pushes to
ensure the elements of the tuple are put in their proper place.

Reviewed By: jano

Differential Revision: D17199454

fbshipit-source-id: 51705d77179bea9a2bb3a62f6e75e3320eb99004