Simplify ClassGetTS interpreter implementation
Summary:
Do not mangle class name together with generics, just to pass it to
lookupClsRef(), which forwarded it to Unit::loadClass(), which unmangled it and
used just the original class name.

Instead, call the Unit::loadClass() with the original class name.

This was the last usage of mangleReifiedName(), i.e. the last place
in the codebase constructing '$$name$$<generics>' reified names.

The next diff will remove all the complexity related to reading these names.

We should probably also kill the reified generics table that tries to make
reified types static, but that's for another diff...

Reviewed By: ricklavoie

Differential Revision: D17200413

fbshipit-source-id: 75d5d5898aca184321036aed2213bf663abff28c