Experiment w/ tracking prov for empty arrays in HHBBC
Summary:
The fundamental problem with doing this is that, before, hhbbc has pretty strict guarantees about what kind of data can be attached to any given type and which type bits that data corresponds to. Now we get to throw that out the window, since we can have a type, e.g. VecN { .... ArrLikePackedN ...} unioned with a VecE { ... empty vec ...} and need to know which arraylike data we need to discard.

Basically the solution I came up with is to introduce a function `project_data` that discards the data of a type if it is invalid for a given trep; when we intersect two types, we first discard data that would be inapplicable for the intersection--this applies equally when we check subtype and couldBe relations (including when we do so as part of the union operation)

After that workaround, the rest is pretty straightforward

Reviewed By: mxw

Differential Revision: D15376674

fbshipit-source-id: 8390452de6f1bc2eee68a41270ee11e785807aa9