Introduce option for not logging $var::meth() as dynamic call
Summary:
Right now expressions looking like `$var::meth()` are logged as dynamic calls. Because in the future thanks to class pointers we don't want them to be considered as such, we don't want to log them as dynamic for codemods working on fixing dynamic calls without pointers. This diff introduces new runtime option to skip logging for expression of this type.

It introduces an additional parameter to calls done using `FCallClsMethod` for whether a call should be logged as dynamic or not.

Reviewed By: ricklavoie

Differential Revision: D16785355

fbshipit-source-id: 982bb2b2b9a2c715b87fecf3dfe4b91c0f4ba7a0