Do not use continuation for obj_get or class_get
Summary:
Method invocations (for instance or static methods) are currently type checked using `obj_get` or `class_get` by passing in a continuation parameter `k` (via `method_get_helper`) that checks the actual call through a particular class.

This is quite contorted, and difficult to comprehend and maintain. Better (if possible) to separate the checking of the member access (`obj_get` or `class_get`) from the checking of the call (using `call`). We get rid of `method_get_helper`, and the use of the continuation `k`.

The trickiest aspect is dealing with nullsafe calls, as these need to propagate nullability from the type of the receiver (e.g. `$x : ?C`) to the type of the result of the method invocation (e.g. if `class C { function foo(string $x):int; }` then `$x->foo("a")` has type `?int`. We do this by propagating nullability to the return type of the *function type* of the method, so in this example we transform `(function(string):int)` to `(function(string):?int)`. But what if we have nothing concrete to work with e.g. `$x` has type `dynamic`, or type `null`? We don't compute a function type in these cases, but let the `call` logic deal with it. For the case of `null`, we now pass in `nullsafe` as a parameter into `call`, so that calls "on null" are supported.

Reviewed By: vassilmladenov

Differential Revision: D16784457

fbshipit-source-id: 5785d982185a5d11f6df31f306413091dc609519