Correct return type in tfun_to_func_details
Summary:
This diff is part of the stack that switches all autocomplete system tests away from the Trie-based autocomplete system to the Sqlite+Local autocomplete system.

Second problem: Some of our existing autocomplete tests assume that we can retrieve the "return type" of a function.  Let's enable our lookup system to produce this return type correctly.

To do this, we make a small modification to tfun_to_func_details and collect the return type out of it.

Reviewed By: 2BitSalute

Differential Revision: D17142600

fbshipit-source-id: 595dafc96b057ebe88f5e6d30dcf318e1e69842e