If fast-facts-parser fails, use FileInfo.t objects
Summary:
This diff is part of the stack that switches all autocomplete system tests away from the Trie-based autocomplete system to the Sqlite+Local autocomplete system.

First problem: Fast facts parser will fail on some incomplete files.  This means if you are in the midst of typing something, FFP will scan the file and sometimes return nothing - which makes it hard to autocomplete symbols in the file you are editing.

To address this, if FFP fails, we will use the FileInfo.t object that we already had.  FileInfo.t objects lack some specific details that FFP produces, so we allow ourselves to use Decl_provider to look up those details.  This will be a low impact to performance since:
* This code path only executes under the following conditions:
  * A file has been modified on disk; and
  * The file could not be parsed by the fast-facts-parser (e.g. it has parsing errors).
* When this occurs, we know that the file that has been modified has been typechecked already.
* Therefore, typechecking has already occurred for these files, and calling Decl_provider won't create extra burden.

Reviewed By: 2BitSalute

Differential Revision: D17142435

fbshipit-source-id: 975fd66feed348209c9eb3ed87e15ef6f1eab4f6