Switch hh_single_type_check to local index only
Summary:
The program `hh_single_type_check` runs using TrieIndex mode as default.  Because this program runs many of the autocomplete tests, this means lots of the autocomplete tests are still using the legacy TrieIndex mode.

This diff switches `hh_single_type_check` to run using the LocalIndex mode.  This mode is appropriate because this program:
* Is intended to simply execute one command and then exit
* Runs only on a single file or a small number of files
* Ideally should not connect to glean or build a full sqlite index

After this diff, symbol index behavior will be unified across all modes, and we can proceed forward to remove the TrieIndex code.

This diff will also enable the usage of the new autocomplete system by the open source build.  However, this diff does not by itself switch all execution to the new system; a final diff after this will switch the default to sqlite mode.

Key differences:
* The old TrieIndex mode operated in a mode where namespaces were considered simply part of the name of a symbol.
* Some of the existing tests assumed that within namespace Foo you would be recommended results that include the name Foo ... but that's not really necessary.
  * case_with_ns.php.exp
  * case_with_prefix.php.exp
* The new symbol index produces results in some cases where the TrieIndex did not.  These cases appear to be scenarios where partially-entered text previously generated no results, but now can find useful options:
  * after_attribute.php.exp
  * case.php.exp
  * dict_keys.php.exp

Reviewed By: 2BitSalute

Differential Revision: D17169459

fbshipit-source-id: 0af714c0200a78f2087dbbf76a253b819856b2e2