Avoid starting hackc until we know we're going to use it
Summary: Minor change - when using the hackc compiler hook, it's not apparent that we'll *ever* need to ensure that the hackc process is start because the hook could never invoke the compiler. This patch makes sure that we only attempt to start the hackc process when we're going to actually compile.

Reviewed By: alexeyt

Differential Revision: D17173677

fbshipit-source-id: 6d0b2f307b1cbcaafe11a44c9547dee5031f99ca