Refactor pBlock in Ocaml lowerer
Summary:
`pStmtUnsafe` wraps a single statement into a list as if a block. Callers of `pStmtUnsafe`  are two kinds, the first kind uses it to lower a block and the second kind uses it lower a statement then concatenate them to a block then do some simple post processing(`pBlock` and `unwrap_extra_block`). For the second kind, `pStmtUnsafe` can be replaced by `pStmt` and remove `List.concat`. For the first kind, two different post processing are unified and simplified to `pBlock` with one additional parameter.

This refactoring simplifies Rust port work and also can improve Ocaml perf as `List.concat`s are removed.

Reviewed By: Matt-Schellhas

Differential Revision: D17214654

fbshipit-source-id: 874cd6a2ea3a0097efc82960509ebca89425991d