Don't emit live translations for a function while it's still being profiled
Summary:
These live translations would be made unreachable once the SrcRec is cleared to
publish the optimized code.  This also avoids an assertion regarding the number
of translations in the SrcRec when we initially try to emit a Profile
translation, but then code.prof fills up by the time the write lease is acquired
and the kind of translation changes to a Live one.

Reviewed By: ricklavoie

Differential Revision: D17233321

fbshipit-source-id: c3994ddf848398feb962e2a7bab7a789c934c927