Skip loops with empty bases
Summary:
Since the length check is hidden in an opaque native helper, we currently generate a full native call for iteration over a static empty base.

This situation comes up a few hundred times in the prod TC, which isn't that much. The common case when it happens is for varargs functions where we rarely pass the extra arguments, like minva / maxva. I don't expect this diff to have any perf impact, but it's nice to clean up the TC.

Reviewed By: mxw

Differential Revision: D17255641

fbshipit-source-id: 3d5fdef9dbc546e4d4b3bac7f88d1cfced9de4b9