Correctly check the variance of generic type parameters appearing at the root of type access in where constraints
Summary:
This diff fixes another instance of the problem described in D16916838: we didn't check the variance of generic type parameters appearing at the root of type access, e.g., of `T` in `T::TA::TB`, because we mistakenly assumed that such types were not allowed, whereas they are but (at the moment) only in `where` constraints.  Consider the following example (added below as a new test case):
```
<?hh
// Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.

interface I {
  const type TThis = this;
}

class BrokenRef<+T as I> {
  public function __construct(private T $value) {}
  public function get<Tu>(): Tu where Tu = T::TThis {
    return $this->value;
  }
  public function set<Tu>(Tu $value): void where Tu = T::TThis {
    $this->value = $value;
  }
}

function swap(BrokenRef<A> $r1, BrokenRef<A> $r2): void {
  $v1 = $r1->get();
  $r1->set($r2->get());
  $r2->set($v1);
}

class A implements I {}

class B extends A {
  public function f(): void {}
}

<<__EntryPoint>>
function test(): void {
  $r1 = new BrokenRef(new A());
  $r2 = new BrokenRef(new B());
  swap($r1, $r2);
  $r2->get()->f();
}
```
The type checker is happy with it but it fatals at run time.  The type checker is failing to see that `T` is declared as covariant but is used as an invariant generic type parameters (because it appears in an equality `where` constraint).

Reviewed By: andrewjkennedy

Differential Revision: D17203313

fbshipit-source-id: 286d5c1dcea09e226751fe1e208a4f1dc6c49a85