Enable synthetic tests with fake "root" folders to work
Summary:
As part of the project to gradually phase out the old "Trie"-based autocomplete, we'll switch our test suite to use LocalIndex.

However, as it turns out, some tests will fake the notion of editing a file and use a phony root value.  Let's allow those tests to proceed even if they are using LocalIndex.

Basically, the tests work as follows:
* Set the root folder to "/"
* Send phony "file edited" messages to hh
* Expect that files edited via this phony process work correctly

Reviewed By: 2BitSalute

Differential Revision: D17228758

fbshipit-source-id: 29c7c82d3784f91319dc3d70b8dced2d825a36a8