Implement documentSymbol for sIDE
Summary:
This diff enables the Document Symbol LSP message in serverless IDE.

Document Symbol is used to locate symbols known in a local file.  It enables quickly searching for names of interest in a kind of "outline" format.

From scuba stats, this message has a lot of lag associated with it, which means we can potentially reduce LSP lag by switching this message to sIDE.

This diff does not change the behavior of the Document Symbol LSP message - it simply takes the existing hh_server behavior and splits it out into sIDE.

Reviewed By: 2BitSalute

Differential Revision: D16837019

fbshipit-source-id: 9c8837d35857866c582eac58b607f57ab767a6ff