Make local index run a full scan of symbols
Summary:
The goal of my diffstack is to switch all the tests to run on the LocalIndex mode, which will ensure that we can remove the old Trie-based index and put all tests on the new autocomplete codepath.

In order to make some tests work, we need to be able to scan the entire project.

To do this, we modify two functions to return the correct values: one indicating whether we should run in partially threaded mode, and one indicating whether we should run a scan at all.

Reviewed By: 2BitSalute

Differential Revision: D17263703

fbshipit-source-id: b07c22ca52810445e5a0575ba3d35b3355594520