Add ocamlpool_generation to ocaml_ffi_mock
Summary:
See {D17111899} for context - looks like it's possible to pull in this variable the same way other functions in this file are, and cause linker errors.

Not 100% sure how it actually works without introducing duplicate definitions, but adding this declaration fixes the issue form linked diff, while not breaking compilation of `hh_parse` (both in `dev` and `dbg`) mode.

Reviewed By: shiqicao

Differential Revision: D17262271

fbshipit-source-id: 26f9b1ea13eb9a10a32461b924fe58d4d75124a8