Rename IntoOcamlRep trait to OcamlRep
Summary: We will need to add more methods to this trait (e.g., we need a convert-to-ocamlrep-by-reference method; see T52744311), so it might make more sense to name the trait OcamlRep instead. As an alternative, we could define several traits, but I don't see a reason why that would be preferable.

Reviewed By: 2BitSalute

Differential Revision: D17284196

fbshipit-source-id: 738110654ff98eb14084aee8d3c2c3f8d37859ff