Do not take &mut Arena in allocation methods
Summary:
In {D16817846}, I changed Arena's allocation methods to borrow the Arena mutably, in order to remove a line of unsafe code in Arena. As I noted in that summary, I didn't have a reason to believe this code was invalid or incorrect, but I wanted to eliminate the unsafe line anyway just in case.

I want to bring it back now. I'd like to convert OCaml values (on the OCaml heap) into ocamlrep values, and I'm not sure how I can do so without making the conversion method re-entrant, and I'm not sure how to make it re-entrant (without paying additional costs at runtime) if it borrows the arena mutably.

Reviewed By: 2BitSalute

Differential Revision: D17263130

fbshipit-source-id: 7a3b7b2944559587c44415f70c8182984b7d9faf