Convert OCaml values into ocamlrep values
Summary: Because the in-memory layout of ocamlrep values and values on the OCaml heap are identical, it is straightforward to copy OCaml-heap values into an ocamlrep Arena. This gives us a simple means of converting OCaml values to Rust values, allowing us to experiment with sharing shallow decls in cachelib using ocamlrep before the Rust shallow-decl-parser is complete.

Reviewed By: 2BitSalute

Differential Revision: D17263129

fbshipit-source-id: 91913152ec95b651dae275515820f3524c3dd9d2