Allow users to configure the Watchman socket path
Summary:
When you run HHVM, some code executes as your user, some as `apache`, and some as `root`. We want to make sure that HHVM always communicates with the Watchman instance that your user is running, no matter the mode.

This diff gets the user who owns your repo and looks in your per-user settings (ours is in `/home/$USER/local/.hphp`) for a `watchman.socket` setting. If none is available, we'll fallback to the `watchman.socket` found in your main config file, and if nothing's *there* we'll fall back to calling `watchman get-sockname`.

Reviewed By: dneiter

Differential Revision: D17066504

fbshipit-source-id: ee0b94fffccc3a8d513e37ed3f36507eefbaa576