Have reified param/return type be consistent with regular ones
Summary:
Make it so that reified param and return type hints behave the same way as literal ones, i.e. `shape('a' => int)` should only check whether it is in fact a shape but not the contents of the shape.

Due to the test changes in this diff, also found a bug where warn and strict were typo'ed in the code.

Reviewed By: alexeyt

Differential Revision: D17269073

fbshipit-source-id: 6914a57374a65e0d128efd0bac950514fe7de671