Disable instance of optimizations for interfaces
Summary:
I have had issues come up with this in previous oncalls.  Last time we papered
over them.

In the long run we probably should add proper interface tracking for types in
the typesystem, but for now it seems like the most sound thing to do is not
perform this unnecessary intersection with an interface type.
-> T36601831

Reviewed By: ricklavoie

Differential Revision: D17290916

fbshipit-source-id: d593fa9e8dd966f6b17cea966166467962297a25