Allow like casts in the typechecker
Summary:
This builds out the typechecker support for like casts. The rules are that a like cast is valid if it is valid without the like, except that types with covariant erased generics and erased generics themselves are legal.

For a cleaner implementation and less confusing error messages, I've removed the distinction between soft reified and erased generics.

Reviewed By: kmeht

Differential Revision: D17223055

fbshipit-source-id: 84da31289027ea94f200b3200768d8ce14c58d20