Unique and persistent attributes for records
Summary:
A record is marked unique in HHBBC if there is only one `RecordInfo` corresponding to a record name (and no classes or type-aliases) and the parent (if any) is also marked unique.
A record is marked persistent if it is unique, it's parent (if any) is marked persistent and the unit it is declared in is persistent and has a persistent pseudomain.

The `RecordInfo` structure currently has minimal information, and there is no `res::Record` yet.

Differential Revision: D16628572

fbshipit-source-id: baaeda7cf04054209c1e349d03f74f514df1778d