Show a more specific error when accessing members on mixed values
Summary:
When users tried to do `$foo->bar` or `$foo->bar()` on a `mixed` value, we previously complained about nullability (because `mixed` is equivalent to `?nonnull`). This isn't helpful, because you can't call methods on `nonnull` values either.

Instead, show a more specific error when we encounter a method call on a top type (`mixed`, `nonnull`). Ensure that we still use the error code that was used before.

Reviewed By: vsiles

Differential Revision: D17283874

fbshipit-source-id: b2f78f0369204cf555f4b470972a9fae71e4b58b