Don't call the hack implementation of idx from the bytecode.
Summary:
The hack implementation of idx doesn't do anything very magical any more, we can do it all in C++.
In the jit, interpOne instead of calling the hack implementation for edge cases.
Replicate the hack implementation in the interpreter.
Add unit test that mqian wrote that demonstrates we don't get dynamic invocation notices when calling idx with unusual keys.

Reviewed By: ricklavoie

Differential Revision: D17313812

fbshipit-source-id: 5634ecc32a9fde924533987df72d61d042c400e0