Fix guard-constraint issue in irgen-builtin
Summary:
During region selection, we emit code given fine-grained live types, but at the same time, we emit guard constraints that specify how closely a type must match the live types for the region to still be valid. For example, the live type for some local may be a specific static string, but the region may be valid for any (static or counted) string.

We only make use of these guard constraints for "old values" that are live on entering the region. For any "new values" computed within the region, we assume that if the guard constraints apply to the old values, they'll imply the required guard constraints for the new ones. This assumption is subtle.

One place where this assumption is very broken is in irgen-builtin. A lot of builtins take mixed inputs and return mixed outputs. We don't constrain parameters with a mixed type-hint, because the native helper will handle them all anyway. However, in many cases, we can generate better IR given certain types for the parameters. When we generate this optimized builtin implementation, we leave the parameters unconstrained, so the "new values" assumption is broken for the output.

Reviewed By: jjgriego

Differential Revision: D17339244

fbshipit-source-id: 16b314b158fe1f0c2031ddc0b54803b8d9673465