Fix autocomplete for lambdas
Summary:
Waleed reported that autocomplete for lambdas doesn't produce the right results.

This was because lambdas started out on the autocomplete pathway for local variables, but ended up in the return pathway for functions.  The return pathway for functions was using a sprintf "%s(%s)" to insert the parameter list and function name... but neither of those two things existed, so the insert text became "()".

To fix this, we'll check before going down the return pathway for functions.

Reviewed By: arxanas

Differential Revision: D17355035

fbshipit-source-id: e6e0c0a1673a50e29215e1eee7db17ff83ee2b7b